UserCollaborative
Added user to user MEA calculation & table MEA that is written to separate files and folders

Calculator.java - corrected meanAbsoluteError method to expect int[][] as realValues (the type of the InitialMatrix)

+++ src/main/Calculator.java
@@ -121,7 +121,7 @@ public class Calculator {
 		return result;
 	}
 	
-	public double meanAbsoluteError(double[] realValues, double[] predicted) {
+	public double meanAbsoluteError(int[] realValues, double[] predicted) {
 		if(realValues.length != predicted.length) {
 			System.err.println("Calculator.java->meanAbsoluteError->(realValues.length != predicted.length)");
 			System.exit(5);
